server/requirements.txt
fastapi
uvicorn[standard]
pydantic
stripe
psycopg2-binary
openai
langchain-openai
langchain-core
python-dotenv